third_party/WebKit/LayoutTests/http/tests/security/XFrameOptions/x-frame-options-deny-expected.txt

http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny.cgi - willSendRequest <NSURLRequest URL http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny.cgi, main document URL http://127.0.0.1:8000/security/XFrameOptions/x-frame-options-deny.html, http method GET> redirectResponse (null)
CONSOLE ERROR: Refused to display 'http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny.cgi' in a frame because it set 'X-Frame-Options' to 'deny'.
http://127.0.0.1:8000/security/XFrameOptions/resources/x-frame-options-deny.cgi - didFinishLoading
data:text/html,chromewebdata - willSendRequest <NSURLRequest URL data:text/html,chromewebdata, main document URL http://127.0.0.1:8000/security/XFrameOptions/x-frame-options-deny.html, http method GET> redirectResponse (null)
data:text/html,chromewebdata - didReceiveResponse <NSURLResponse data:text/html,chromewebdata, http status code 200>
data:text/html,chromewebdata - didFinishLoading
CONSOLE MESSAGE: line 14: PASS: Access to contentWindow.location.href threw an exception.
There should be no content in the iframe below



--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
